clean: rm nonfunctional 'shareable link'
While working on documentation per https://github.com/StatCan/daaas/issues/113, I discovered that the "get shareable link" option in the Access Port tool is nonfunctional. I have removed this option so as to not introduce any unnecessary confusion.

With how our Kubeflow RBAC is set up, whether or not the generated port access URL is the "shareable" version, logged-in contributors to the namespace can access it, while others cannot.