feat(security): checksum RStudio
This checksum was missed in the first pass, seems to be the only one though.

Note: while here I took a look at updating RStudio, but changing either the Ubuntu release (to bionic given that this is an Ubuntu 18 base) or the version number results in crashes. I figure this would need to be a more complex procedure in conjunction with r-runtime updates.