fix(filebrowser): Loading hotfix
I thought this part was unnecessary, but it actually breaks Filebrowser with a `Please provide a valid API token via token get parameter.` down the line. I'm not 100% sure why this is effect is very delayed, but I think it might have something to do with the token being saved as a cookie that will eventually expire. To the extent of my experimentation, I am not aware of this granting any undue privileges, given that access is ultimately controlled by authentication to Kubeflow and RBAC (e.g. namespace collaborators).

Discovered during work on https://github.com/StatCan/kubeflow-containers/issues/38 but ultimately unrelated