clean: Fix indent, netdata config near install
Fixes an indent in the netdata install and moves the netdata configuration next to it. The netdata configuration is presently commented out (and the folder in question no longer exists) as we are going for default configuration for this version. The comment is left as a reference (and now in a more appropriate location) should that change in the future.